+++ src/main/kotlin/com/github/insanusmokrassar/IObjectKRealisations/Example.kt
@@ -1,0 +1,32 @@
+package com.github.insanusmokrassar.IObjectKRealisations
+
+import com.github.insanusmokrassar.IObjectK.interfaces.IObject
+import org.json.JSONObject
+
+fun main(args: Array<String>) {
+    val jsonObjectExample = JSONObject()
+    jsonObjectExample.put("Hello", "World")
+    jsonObjectExample.put("InnerObject", JSONObject("{\"Inner hello\":\"World\"}"))
+    val jsonIObject = JSONIObject(jsonObjectExample).observable()
+    jsonIObject.observable.subscribe({
+        it.second ?. let { _ ->
+            println("Key: ${it.first}, Value: ${it.second}")
+        } ?: println("Value by key '${it.first}' deleted")
+    })
+    println(jsonIObject.toString())
+    println(jsonIObject.get<IObject<Any>>("InnerObject").toString())
+
+    jsonIObject.putAll(
+            mapOf(
+                    Pair("Example", "Example"),
+                    Pair("Example", "Example2"),
+                    Pair("Example", "Example3")
+            )
+    )
+
+    jsonIObject.put("Example", "Example")
+    jsonIObject.put("Example", 2)
+    jsonIObject.put("Example", true)
+
+    jsonIObject.remove("Example")
+}